agno_tests/pyproject.toml
[project]
name = "agno-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.5",
    "google-genai>=1.53.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
]
